Ya funciona correctamente con el control de excepciones el modulo de proveedores

+++ src/main/java/com/example/sistemaComplejoDeportivo/service/ProveedorService.java
@@ -1,6 +1,7 @@
 package com.example.sistemaComplejoDeportivo.service;
 
 import com.example.sistemaComplejoDeportivo.model.Proveedor;
+import com.example.sistemaComplejoDeportivo.repository.InventarioRepository;
 import com.example.sistemaComplejoDeportivo.repository.ProveedorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,8 @@ public class ProveedorService {
 
     @Autowired
     private ProveedorRepository proveedorRepository;
+    @Autowired
+    private InventarioRepository inventarioRepository;
 
     public List<Proveedor> obtenerTodosLosProveedores() {
         return proveedorRepository.findAll();
@@ -26,7 +29,11 @@ public class ProveedorService {
         return proveedorRepository.save(proveedor);
     }
 
-    public void eliminarProveedor(Integer id) {
+    public void eliminarProveedor(Integer id) throws Exception {
+        if (inventarioRepository.existsByProveedor_IdProveedor(id)) {
+            throw new Exception("No se puede eliminar el proveedor porque tiene productos en el inventario.");
+        }
         proveedorRepository.deleteById(id);
     }
+
 }
